I moved the login to the right place
