+++ .babelrc
@@ -1,3 +1,8 @@
 {
-"presets": ["react-native"]
+ "presets": [
+   "react-native"
+ ],
+ "plugins": [
+    "transform-function-bind"
+  ]
 }
