+++ index.android.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { AppRegistry } from 'react-native'
 
-import App from './src/containers/App'
+import App from './src/scenes/root/App'
 
 AppRegistry.registerComponent('Timekeeper', () => App)
