+++ package.json
@@ -21,6 +21,7 @@
   "devDependencies": {
     "babel-jest": "18.0.0",
     "babel-preset-react-native": "1.9.1",
+    "babel-plugin-transform-function-bind": "^6.22.0",
     "jest": "18.1.0",
     "react-test-renderer": "15.4.2"
   },
